requirements.txt
torch>=0.4.1 # for out of date CUDA versions, go to https://pytorch.org/ to install torch
torchvision
numpy
scikit-image
matplotlib
cloud-volume==0.45.0
fpzip==1.1.2
h5py
moviepy
neuroglancer
visdom
pathos
pyshorteners
argcomplete
pandas
opencv-python
setuptools>=34.0.0
pyyaml
task-queue
imageio==2.4.1
